+++ resources/terraform/testnet/digital-ocean/production.tfvars
@@ -5,7 +5,7 @@ bootstrap_droplet_image_id = 157362431
 node_droplet_size = "s-2vcpu-4gb"
 node_vm_count = 79
 node_droplet_image_id = 157362431
-private_node_vm_count = 1
+private_node_vm_count = 0
 uploader_droplet_size = "s-2vcpu-4gb"
 uploader_vm_count = 5
 uploader_droplet_image_id = 157362431
